Fix: API endpoint /api/routes alters the specified coordinates
Bug: when sending a request to API resource
/api/routes/{seaportuid}/{lat}:{lon} with a negative longitude
(for example -7.9251411000), the client gets a response where
the given longitude changed to a positive value (172.0748589).
But the specified coordinates should remain unchanged.

+++ src/main/java/net/contargo/iris/route/RouteInformation.java
@@ -3,6 +3,8 @@ package net.contargo.iris.route;
 import net.contargo.iris.GeoLocation;
 import net.contargo.iris.container.ContainerType;
 
+import java.math.BigDecimal;
+
 import java.util.Objects;
 
 
@@ -22,7 +24,7 @@ public class RouteInformation {
     public RouteInformation(ContainerType containerType, RouteCombo combo, double lat, double lon, boolean isRoundTrip,
         boolean isImport) {
 
-        this.destination = new GeoLocation(lat, lon);
+        this.destination = new GeoLocation(BigDecimal.valueOf(lat), BigDecimal.valueOf(lon));
         this.containerType = containerType;
         this.routeCombo = combo;
         this.product = RouteProduct.fromIsRoundtrip(isRoundTrip);
